move stuff to submit controller, tokenize, other stuff. Not 100% yet.
